% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSimilarity.R
\name{calculateSimilarity}
\alias{calculateSimilarity}
\title{Calculate hamming distance between samples}
\usage{
calculateSimilarity(hla_data)
}
\arguments{
\item{hla_data}{data frame with allele data}
}
\value{
TODO
}
\description{
Calculate hamming distance between samples
}
\examples{

## Load example dataset
data(example_calculateSimilarity)

## Calculate similarity metrics
#calculateSimilarity(example_calculateSimilarity)

}
\author{
Santiago Medina, Nissim Ranade
}
