% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAClustRView.R
\docType{data}
\name{demoHLADataset}
\alias{demoHLADataset}
\title{The HLA typing information from multiple samples stored in a
\code{HLADataset} format (for demo purpose).}
\format{a \code{list} of class \code{HLADataset}
containing the following elements:
\itemize{
\item \code{data} a \code{tibble} object containing the HLA typing
information for all samples. The columns are:
\itemize{
\item \code{SampleName} a \code{character} string that represent the
name of the sample.
\item \code{AllelName} a \code{character} string that represent the
name of the allele (1 or 2).
\item \code{GeneName} a \code{character} string that represent the
name of the HLA gene.
\item \code{AlleleGroup} a \code{character} string that represent the
section identifying the subtype group.
\item \code{Protein} a \code{character} string that represent the
section identifying the nucleotide substitutions group.
\item \code{SynSubst} a \code{character} string that represent the
section identifying the synonymous nucleotide substitutions group.
\item \code{NonCoding} a \code{character} string that represent the
section identifying the non-coding polymorphisms group.
\item \code{Suffix} a \code{character} string that represent the
suffix of the HLA typing or \code{NA}.
}}}
\usage{
data(demoHLADataset)
}
\value{
a \code{list} of class \code{HLADataset}
containing the following elements:
\itemize{
\item \code{data} a \code{tibble} object containing the HLA typing
information for all samples. The columns are:
\itemize{
\item \code{SampleName} a \code{character} string that represent the
name of the sample.
\item \code{AllelName} a \code{character} string that represent the
name of the allele (1 or 2).
\item \code{GeneName} a \code{character} string that represent the
name of the HLA gene.
\item \code{AlleleGroup} a \code{character} string that represent the
section identifying the subtype group.
\item \code{Protein} a \code{character} string that represent the
section identifying the nucleotide substitutions group.
\item \code{SynSubst} a \code{character} string that represent the
section identifying the synonymous nucleotide substitutions group.
\item \code{NonCoding} a \code{character} string that represent the
section identifying the non-coding polymorphisms group.
\item \code{Suffix} a \code{character} string that represent the
suffix of the HLA typing or \code{NA}.
}}
}
\description{
This dataset can be used to test \code{calculateHamming} and
TODO(AUTRE METRIC) functions.
}
\examples{

## Load example dataset
data(demoHLADataset)

## Calculate Hamming distance metric
calculateHamming(demoHLADataset)


}
\keyword{datasets}
