% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cluster.R
\name{draw_cluster}
\alias{draw_cluster}
\title{Draw cluster}
\usage{
draw_cluster(matrixDistance)
}
\arguments{
\item{matrixDistance}{distance matrix}
}
\value{
TODO
}
\description{
TODO
}
\examples{

##TODO

}
\author{
Santiago Medina, Pascal Belleau
}
