% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cluster.R
\name{draw_dendogram}
\alias{draw_dendogram}
\title{Draw dendrogram}
\usage{
draw_dendogram(hlaMetric)
}
\arguments{
\item{hlaMetric}{an object of class \code{hlaMetric} that contain the
distance between samples.}
}
\value{
TODO
}
\description{
Create and draw a dendrogram using the
metric distances.
}
\examples{

## Load example dataset
data(demoHLADataset)

## Calculate Hamming distance metric
hammingMetric <- calculateHamming(demoHLADataset)

## Create dendogram
draw_dendogram(hammingMetric)

}
\author{
Santiago Medina, Astrid Deschenes, Pascal Belleau
}
