% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cluster.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Draw sample-to-sample heatmap}
\usage{
draw_heatmap(hlaMetric, ...)
}
\arguments{
\item{hlaMetric}{an object of class \code{hlaMetric} that contain the
distance between samples.}

\item{\ldots}{arguments passed to the \code{Heatmap} method}
}
\value{
A \code{Heatmap-class} object.
}
\description{
Create and draw a sample-to-sample heatmap using the
metric distances.
}
\examples{

## Load example dataset
data(demoHLADataset)

## Calculate Hamming distance metric
hammingMetric <- calculateHamming(demoHLADataset)

## Create sample-to-sample heatmap
draw_heatmap(hammingMetric)

}
\author{
Astrid Deschênes, Pascal Belleau
}
