% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_calculateSimilarity}
\alias{example_calculateSimilarity}
\title{Example toy data for function example_calculateSimilarity}
\format{A data frame with 26 rows and 4 variables:
\describe{
  \item{SampleName}{sample id}
  \item{GeneName}{hla gene id}
  \item{AlleleName}{allele information, 1 or 2}
  \item{AlleleGroup}{allele group data}
  ...
}}
\usage{
example_calculateSimilarity
}
\description{
Toy example data of input for function calculateSimilarity
}
\keyword{datasets}
