% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sample_pair_data}
\alias{example_sample_pair_data}
\title{Example toy data for function sample_pair_distance}
\format{A data frame with 10 rows and 4 variables:
\describe{
  \item{sample}{sample id}
  \item{HLAgene}{hla gene id}
  \item{Allele}{allele information, 1 or 2}
  \item{Digit1}{digit 1 data}
  ...
}}
\usage{
example_sample_pair_data
}
\description{
A dataset containing the allele and digit1 information for
genes in a pair of samples
}
\keyword{datasets}
