% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{extractRef}
\alias{extractRef}
\title{Process the line contening the ref seq}
\usage{
extractRef(seq, startPos)
}
\arguments{
\item{seq}{a \code{string} contening the reference sequence.}

\item{startPos}{a \code{integer} corresponding to the position in the string
where the sequence is starting}
}
\value{
a \code{list} with two fields. The first is \code{refSeq} which is
containing the sequence of the reference. The second is seqDiff is a
\code{string} representing the alignment format for the current sequence.
}
\description{
TODO
}
\examples{

## TODO

}
\author{
Pascal Belleau
}
\keyword{internal}
