% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{extractTyping}
\alias{extractTyping}
\title{Extract the typing string from line contening the alignment}
\usage{
extractTyping(seq, endPos)
}
\arguments{
\item{seq}{A \code{string} contening the sequence.}
}
\value{
TODO
}
\description{

}
\examples{

}
\author{
Pascal Belleau
}
\keyword{internal}
