% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAdbTools.R
\name{getSeqCMP}
\alias{getSeqCMP}
\title{Get the sequence from the type}
\usage{
getSeqCMP(HLAInfo, regionExt, typeS1, typeS2)
}
\arguments{
\item{HLAInfo}{TODO}

\item{regionExt}{TODO}

\item{typeS1}{Sample 1 TODO}

\item{typeS2}{Sample 2 TODO}
}
\value{
An object of class TODO
}
\description{
Get from the object HLAdb two sequences and the reference
for a region
}
\details{
TODO
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschenes
}
