% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAdbTools.R
\name{getSubSeq}
\alias{getSubSeq}
\title{Get the subsequence}
\usage{
getSubSeq(seq, posInit, regionExt)
}
\arguments{
\item{seq}{TODO}

\item{posInit}{TODO}

\item{regionExt}{TODO}
}
\value{
An object of class TODO
}
\description{
Get from the object HLAdb two sequences and the reference
for a region
}
\details{
TODO
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschenes
}
