% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{getTypingPos}
\alias{getTypingPos}
\title{TODO}
\usage{
getTypingPos(seqProcess, curTyping)
}
\arguments{
\item{curTyping}{A \code{vector} of \code{character} strings defining
the HLA type. The \code{vector} should have 6 entries.}

\item{seq}{A \code{character} string containing the sequence.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO

}
\author{
Pascal Belleau
}
\keyword{internal}
