% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSimilarity.R
\name{hamming_distance_digit1}
\alias{hamming_distance_digit1}
\title{Compute the hamming distance between two alleles}
\usage{
hamming_distance_digit1(allele)
}
\arguments{
\item{allele}{tibble with 3 columns: *sample*, *Allele*, *Digit1*
(see example)}
}
\value{
One row tibble with minimum hamming distance &
logical value column indicating whether same allele was used If the distance is the
same for both then NA is returned
}
\description{
This function computes the hamming distance for two samples and two alleles,
 the hamming distance is given by
\eqn{min(s1A1 != s2A1 + s1A2 != s2A2, s1A1 != s2A2 + s1A2 != s2A1)}
}
\examples{
d <-
tibble::tribble(
   ~sample, ~Allele, ~Digit1,
   "s1", 1, 1,
   "s1", 2, 3,
   "s2", 1, 1,
   "s2", 2, 5,
)
hamming_distance_digit1(d)
}
\author{
Santiago Medina, Nissim Ranade
}
