% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAClustRView.R
\docType{package}
\name{HLAClustRView-package}
\alias{HLAClustRView-package}
\alias{HLAClustRView}
\title{HLAClustRView: TODO}
\description{
This package is TODO
}
\author{
Nissim Ranade,
Santiago Medina,
Adewunmi Adelaja,
Astrid Deschenes and
Pascal Belleau

Maintainer:
Pascal Belleau <belleau@cshl.edu>
}
\keyword{package}
