% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.HLADataset.R
\name{print.HLADataset}
\alias{print.HLADataset}
\title{Print a \code{HLADataset} object}
\usage{
\method{print}{HLADataset}(x, ...)
}
\arguments{
\item{x}{the output object from \code{readHLADataset}
function to be printed}

\item{\ldots}{arguments passed to or from other methods}
}
\value{
an object of class
\code{HLADataset}
}
\description{
Print a \code{HLADataset} object
}
\examples{

## Load dataset
## TODO

}
