% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAClustRView.R
\docType{data}
\name{hladb}
\alias{hladb}
\title{This dataset contains the information from the HLA database.}
\format{TODO}
\usage{
data(hladb)
}
\value{
TODO
}
\description{
This dataset contains the information from the HLA database.
}
\details{
See \url{http://hla.alleles.org/alleles/text_index.html}
}
\examples{

## TODO

}
\keyword{datasets}
