% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAClustRView.R
\docType{data}
\name{hladb_protein_3.35.0}
\alias{hladb_protein_3.35.0}
\title{This dataset contains the protein information from the HLA
database version 3.35.0.}
\format{An object of class \code{HLAdb} with 3 entries. The entries are:
\itemize{
\item \code{refSeq} A \code{list} of \code{character} string that
represent reference sequences; one sequence
per HLA gene.
\item \code{posInit} A \code{list} of \code{integer};
one starting position per HLA gene.
\item \code{HLAAlignment} A \code{data.table} containing the information
for each allele of each HLA gene.
}}
\usage{
data(hladb_protein_3.35.0)
}
\value{
An object of class \code{HLAdb} with 3 entries. The entries are:
\itemize{
\item \code{refSeq} A \code{list} of \code{character} string that
represent reference sequences; one sequence
per HLA gene.
\item \code{posInit} A \code{list} of \code{integer};
one starting position per HLA gene.
\item \code{HLAAlignment} A \code{data.table} containing the information
for each allele of each HLA gene.
}
}
\description{
This dataset contains the protein information from the HLA
database version 3.35.0.
}
\details{
See \url{http://hla.alleles.org/alleles/text_index.html}
}
\examples{

## TODO

}
\keyword{datasets}
