% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.HLAMetric.R
\name{print.HLAMetric}
\alias{print.HLAMetric}
\title{Print a \code{HLAMetric} object}
\usage{
\method{print}{HLAMetric}(x, ...)
}
\arguments{
\item{x}{the output object from \code{calculateHamming}
function to be printed}

\item{\ldots}{arguments passed to or from other methods}
}
\value{
an object of class
\code{HLAMetric}
}
\description{
Print a \code{HLAMetric} object
}
\examples{

## Load example dataset
data(demoHLADataset)

## Calculate Hamming distance metric
metric <- calculateHamming(demoHLADataset)

## Print information about metric
print(metric)

}
