% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{loadHLADb}
\alias{loadHLADb}
\title{load HLA Database alignment files for a specific alignment type}
\usage{
loadHLADb(hlaDbFile)
}
\arguments{
\item{hlaDbFile}{A \code{character} string, the path to the rds file from
parseHLADbAlignment.
#'}
}
\value{
An object of class \code{HLAdb} with 3 entries. The entries are:
\itemize{
\item \code{refSeq} A \code{list} of \code{character} string that
represent reference sequences; one sequence
per HLA gene.
\item \code{posInit} A \code{list} of \code{integer};
one starting position per HLA gene.
\item \code{HLAAlignment} A \code{data.table} containing the information
for each allele of each HLA gene.
}
}
\description{
load HLA Database parse by parseHLADbAlignment
}
\details{
See \url{http://hla.alleles.org/alleles/text_index.html}
}
\examples{

## Get path where some HLA database files are stored
##directory <- system.file("extdata", package = "HLAClustRView")

## Parse HLA database files of protein type
##HLAInfo <- loadHLADb(hlaDbFile="")

## Show reference sequences
##HLAInfo$refSeq

}
\author{
Pascal Belleau, Astrid Deschenes
}
