% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSimilarity.R
\name{make_distance_matrix}
\alias{make_distance_matrix}
\title{Convert distance table to distance matrix}
\usage{
make_distance_matrix(outMet)
}
\arguments{
\item{outMet}{a \code{tibble} object with Hamming distance for sample
pairs \code{calculateHamming}}
}
\value{
a \code{matrix} containing the distance for sample pairs.
}
\description{
This function takes the output of \code{calculateHamming}
function and returns a distance matrix to use for clustering.
}
\examples{

## Load example dataset
data(demoHLADataset)

## Calculate Hamming distance metric
hamming <- calculateHamming(demoHLADataset)

## Calculate Hamming distance matrix
HLAClustRView:::make_distance_matrix(hamming)

}
\author{
Santiago Medina, Pascal Belleau
}
\keyword{internal}
