% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{parseAlignment}
\alias{parseAlignment}
\title{Pre-process HLA Database alignment file.}
\usage{
parseAlignment(fileName)
}
\arguments{
\item{fileName}{a \code{character} string, the name of the alignment file
from \url{http://hla.alleles.org/alleles/text_index.html}}
}
\value{
an object of class \code{HLAGene}
}
\description{
Extract information from one HLA alignment file.
}
\details{
TODO
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
