% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAdbTools.R
\name{parseSubMatrix}
\alias{parseSubMatrix}
\title{load matrix of substitution}
\usage{
parseSubMatrix(fileName)
}
\arguments{
\item{fileName}{TODO}
}
\value{
a matrix
}
\description{
Load matrix BLOSOM or PAM
from ftp://ftp.ncbi.nih.gov/blast/matrices/
}
\details{
TODO
}
\examples{



}
\author{
Pascal Belleau, Astrid Deschenes
}
