% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{reduceTypingPos}
\alias{reduceTypingPos}
\title{Verify if all the sequence are the same for the
typing position}
\usage{
reduceTypingPos(seqProcess, typingPos)
}
\arguments{
\item{TODO}{}
}
\value{
TODO
}
\description{
TODO
}
\details{
TODO
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
