% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSimilarity.R
\name{sample_pair_distance}
\alias{sample_pair_distance}
\title{Sample Pair Distance}
\usage{
sample_pair_distance(sample_pair_data)
}
\arguments{
\item{sample_pair_data}{a tible with columns HLAgene (genes), Allele (alleles),
Digit1 (digit 1) and sample (see example data)}
}
\value{
a tibble with one row and a column for the total hamming distance
and n columns more where n is the number of genes
}
\description{
Computes the Hamming Distance for a pair of samples and then aggregates the
sum over the genes keeping information of which allele was used for the distance
}
\examples{
data("example_sample_pair_data") # example data
sample_pair_distance(example_sample_pair_data)
}
\author{
Santiago Medina, Nissim Ranade
}
