% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTyping.R
\name{splitTyping}
\alias{splitTyping}
\title{Split a HLA allele name into a matrix of fields}
\usage{
splitTyping(aChar)
}
\arguments{
\item{aChar}{a \code{character} array containing one or more well-formed
HLA alleles,
see http://hla.alleles.org/nomenclature/naming.html}
}
\value{
a \code{matrix} of fields with 6 columns and a number of rows
corresponding to the length of the \code{aChar} array.
}
\description{
This function converts a well-formed HLA allele name to a
matrix of fields.
}
\examples{

## Split one HLA typing into an array of entries
output <- HLAClustRView:::splitTyping("HLA-DRB1*13:01:01:02")

print(output)

## Split one HLA typing into an array of entries
output <- HLAClustRView:::splitTyping(c("G*01:04:01:02", "DQA1*01:01:01:03"))

print(output)

}
\author{
Adewunmi Adelaja
}
\keyword{internal}
