% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSimilarity.R
\name{calculateHamming}
\alias{calculateHamming}
\title{calculate hamming distance between samples}
\usage{
calculateHamming(hla_data)
}
\arguments{
\item{hla_data}{data frame with allele data}
}
\value{
a \code{tibble} object containing the Hamming distance values
between each possible pair of samples. TODO
}
\description{
Takes the parsed hla database and computes the hamming distance
for each pair of samples
}
\examples{

## Load example dataset
data(example_calculateSimilarity)

## Calculate hamming distance metric
calculateHamming(example_calculateSimilarity)

}
\author{
Santiago Medina, Nissim Ranade
}
