% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSimilarity.R
\name{calculateHamming}
\alias{calculateHamming}
\title{Calculate Hamming distance between all samples}
\usage{
calculateHamming(hla_data)
}
\arguments{
\item{hla_data}{a \code{list} of class \code{HLADataset} containing a
\code{tibble} object with the HLA typing information for all samples. At
least 2 samples must be present be able to calculate the metric.}
}
\value{
a \code{list} of class \code{HLAMetric}
containing the following elements:
\itemize{
\item \code{metric} a \code{character} string containing the name of the
metric "Hamming Distance".
\item \code{dist} a \code{matrix} that contains the Hamming distance between
samples.
}
}
\description{
Takes an object containing HLA typing information for all
samples and computes the Hamming distance
for each pair of samples.
}
\details{
The Hamming distance is given by
\eqn{min(s1A1 != s2A1 + s1A2 != s2A2, s1A1 != s2A2 + s1A2 != s2A1)} where
\code{s1} and \code{s2} are two samples and \code{A1} and \code{A2} are
there alleles.
}
\examples{

## Load example dataset
data(demoHLADataset)

## Calculate Hamming distance metric
calculateHamming(demoHLADataset)

}
\author{
Santiago Medina, Nissim Ranade
}
