% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAClustRView.R
\docType{data}
\name{example_sample_pair_data}
\alias{example_sample_pair_data}
\title{This dataset contains the allele and digit1 information for
HLA genes in one pair of samples.}
\format{A \code{tibble} object with 10 rows and 4 variables:
\itemize{
    \item{SampleName}{sample id}
    \item{GeneName}{hla gene id}
    \item{AlleleName}{allele information, 1 or 2}
    \item{AlleleGroup}{allele group data}
}}
\usage{
data(example_sample_pair_data)
}
\value{
TODO
}
\description{
This dataset can be used to test the \code{sample_pair_distance} function.
}
\examples{

## TODO

}
\keyword{datasets}
