% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{extractRef}
\alias{extractRef}
\title{Process the line contening the reference sequence}
\usage{
extractRef(seq, startPos)
}
\arguments{
\item{seq}{A \code{character} string contening the reference sequence.}

\item{startPos}{A \code{integer} corresponding to the position in the string
where the sequence is starting.}
}
\value{
A \code{list} with two entries. The first is \code{refSeq} which is
containing the sequence of the reference. The second is seqDiff is a
\code{string} representing the alignment format for the current sequence.
}
\description{
Extract the information about the reference sequence.
}
\examples{

## String with the sequence information
sequence <- " DOB*01:01:01:01       MGSGWV PWVVALLVNL TRLDSSMTQG TDSPEDFVIQ"

## Extract sequence information
HLAClustRView:::extractRef(seq=sequence, startPos=20)

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
