% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{extractSeq}
\alias{extractSeq}
\title{Process the line contening the alignment to the reference}
\usage{
extractSeq(seq, startPos)
}
\arguments{
\item{seq}{A \code{string} contening the reference sequence.}

\item{startPos}{A \code{integer} corresponding to the position in the string
where the sequence is starting}
}
\value{
TODO
}
\description{

}
\examples{

}
\author{
Pascal Belleau
}
\keyword{internal}
