% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{extractTyping}
\alias{extractTyping}
\title{Extract the HLA typing string from one line of aligment}
\usage{
extractTyping(seq, endPos)
}
\arguments{
\item{seq}{a \code{character} string contening the sequence}

\item{endPos}{a \code{integer} fixing the end of the introduction
section of the sequence}
}
\value{
a \code{character} string containing the HLA type of a empty
\code{character} string
}
\description{
Extract the HLA typing string from a character string
containing one line of alignment
}
\examples{

## One line of aligment
sequence <- " DOB*01:01:01:01       MGSGWV PWVVALLVNL TRLDSSMTQG"

## Position of the end of the introduction section
endPosition <- 20

## Extract HLA type
HLAClustRView:::extractTyping(seq=sequence, endPos=endPosition)


}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
