% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAdbTools.R
\name{getSeqCMP}
\alias{getSeqCMP}
\title{Get the sequence from 2 allele types from the same HLA gene.}
\usage{
getSeqCMP(HLAInfo, regionExt, typeS1, typeS2)
}
\arguments{
\item{HLAInfo}{An object of class \code{HLAdb} contain information from
HLA database.}

\item{regionExt}{TODO}

\item{typeS1}{A line of the tibble data from the object data return readHLADataset}

\item{typeS2}{A line of the tibble data from the object data return readHLADataset
HLA allele.}
}
\value{
An object of class TODO
}
\description{
Extract from the \code{HLAdb} object, two sequences and
the reference for a specific region related to 2 specific allele types.
}
\details{
TODO
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschenes
}
