% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{getTypingPos}
\alias{getTypingPos}
\title{TODO}
\usage{
getTypingPos(seqProcess, curTyping)
}
\arguments{
\item{curTyping}{TODO}

\item{seq}{a \code{character} string contening the sequence.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## TODO

}
\author{
Pascal Belleau
}
\keyword{internal}
