% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{getTypingPos}
\alias{getTypingPos}
\title{TODO}
\usage{
getTypingPos(seqProcess, curTyping)
}
\arguments{
\item{seq}{A \code{string} contening the sequence.}
}
\value{
TODO
}
\description{

}
\examples{

}
\author{
Pascal Belleau
}
\keyword{internal}
