% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSimilarity.R
\name{hamming_distance_digit1}
\alias{hamming_distance_digit1}
\title{Compute the Hamming distance between two samples}
\usage{
hamming_distance_digit1(allele)
}
\arguments{
\item{allele}{a \code{tibble} object with 3 mandatory columns:
"SampleName", "AlleleName" and "AlleleGroup".}
}
\value{
a \code{tibble} object with one row containing the minimum
Hamming distance and a \code{logical} column indicating whether
same allele was used. If the
distance is the same for both, then \code{NA} is returned.
}
\description{
Computes the Hamming distance for two samples using both
alleles. See details section for more information about the Hamming
distance.
}
\details{
The Hamming distance is given by
\eqn{min(s1A1 != s2A1 + s1A2 != s2A2, s1A1 != s2A2 + s1A2 != s2A1)} where
\code{s1} and \code{s2} are two samples and \code{A1} and \code{A2} are
there alleles.
}
\examples{

## Create a demo dataset
demoData <- data.frame(SampleName=c("s1", "s1", "s2", "s2"),
        AlleleName=c(1, 2, 1, 2), AlleleGroup=c(1, 3, 1, 5))

## Transform to tibble object
demoDataTibble <- dplyr::as_tibble(demoData)

## Calculate the Hamming distance on the demo dataset
HLAClustRView:::hamming_distance_digit1(demoDataTibble)

}
\author{
Santiago Medina, Nissim Ranade
}
\keyword{internal}
