% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAClustRView.R
\docType{data}
\name{hladb}
\alias{hladb}
\title{This dataset can be used TODO.}
\format{TODO}
\usage{
TODO
}
\value{
TODO
}
\description{
This dataset can be used TODO.
}
\examples{

## TODO

}
\keyword{datasets}
