% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cluster.R
\name{make_distance_matrix}
\alias{make_distance_matrix}
\title{Convert distance table to distance matrix}
\usage{
make_distance_matrix(outMet)
}
\arguments{
\item{outMet}{tibble with hamming distance for sample
pairs \code{calculateHamming}}
}
\value{
TODO
}
\description{
This function takes the output of \code{calculateHamming}
function and returns a distance matrix to use for clustering
}
\examples{

##TODO

}
\author{
Santiago Medina, Pascal Belleau
}
