% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSimilarity.R
\name{makeDistanceMatrix}
\alias{makeDistanceMatrix}
\title{Convert distance table to distance matrix}
\usage{
makeDistanceMatrix(outMet)
}
\arguments{
\item{outMet}{a \code{tibble} object containing distance metric for sample
pairs.}
}
\value{
a \code{matrix} containing the distance for sample pairs.
}
\description{
This function takes a distance table under the form of a
\code{tibble} object and transforms it into a distance matrix
ready for clustering.
}
\examples{

## Tibble object containing a distance table
pairedData <- data.frame(SampleName1=c("ERR053", "ERR053", "ERR465"),
    SampleName2=c("ERR465", "ERR040", "ERR040"),
    HammingDistance = c(17, 18, 19))

tibbleData <- dplyr::as_tibble(pairedData)

## Calculate distance matrix
HLAClustRView:::makeDistanceMatrix(tibbleData)

}
\author{
Santiago Medina, Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
