% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSimilarity.R
\name{parse_hla_data}
\alias{parse_hla_data}
\title{Parse hladb}
\usage{
parse_hla_data(hladb)
}
\arguments{
\item{hladb}{hla database the output of function \code{parseHLADb}}
}
\description{
Converts object to tibble, removes samples with missing values
}
