% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{parseAlignment}
\alias{parseAlignment}
\title{Pre-process HLA alignment file.}
\usage{
parseAlignment(fileName)
}
\arguments{
\item{fileName}{the name of the alignment file from
\url{http://hla.alleles.org/alleles/text_index.html}}
}
\value{
Return a class \code{HLAGene}
}
\description{

}
\details{

}
\examples{

}
\author{
Pascal Belleau
}
