% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{parseHLADbAlignment}
\alias{parseHLADbAlignment}
\title{Parse HLA Database alignment files for a specific alignment type}
\usage{
parseHLADbAlignment(hlaDbPath, seqType = c("prot", "nuc", "gen"))
}
\arguments{
\item{hlaDbPath}{A \code{character} string, the path to the directory of
the alignment files from
\url{http://hla.alleles.org/alleles/text_index.html}. The directory must
exist and contain at least one HLA alignment file.}

\item{seqType}{A \code{character} string, the sequence type of the file to
parse. The choices are: "nuc" for CDS sequence, "gen" for genomic, and
"prot" for protein. Default: "prot".}
}
\value{
An object of class \code{HLAdb} with 3 entries. The entries are:
\itemize{
\item \code{refSeq} A \code{list} of \code{character} string that
represent reference sequences; one sequence
per HLA gene.
\item \code{posInit} A \code{list} of \code{integer};
one starting position per HLA gene.
\item \code{HLAAlignment} A \code{data.table} containing the information
for each allele of each HLA gene.
}
}
\description{
Parse HLA Database alignment files that are present in a
directory, as specified by input, an
generate a class object that can be used in further analysis. The function
only parse one type of aligment file at the time. There is 3 types of
alignment files that can be parsed: CDS sequence, genomic and protein.

Beware that the names of the alignment files should not be changed as
the name is used to identify the gene that is currently parsed.
}
\details{
See \url{http://hla.alleles.org/alleles/text_index.html}
}
\examples{

## Get path where some HLA database files are stored
directory <- system.file("extdata", package = "HLAClustRView")

## Parse HLA database files of protein type
HLAInfo <- parseHLADbAlignment(hlaDbPath=directory, seqType="prot")

## Show reference sequences
HLAInfo$refSeq

}
\author{
Pascal Belleau, Astrid Deschenes
}
