% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHLAdb.R
\name{parseHLADbAlignment}
\alias{parseHLADbAlignment}
\title{Process the HLA alignement type for type (nuc, prot, gen) in a
directory.}
\usage{
parseHLADbAlignment(hlaDbPath, seqType = c("nuc", "gen", "prot"))
}
\arguments{
\item{hlaDbPath}{the path to the directory of the alignment file from
\url{http://hla.alleles.org/alleles/text_index.html}}

\item{seqType}{Sequence type nuc for CDS sequence, gen for genomic, and
prot for protein}
}
\value{
Return a class \code{HLAdb}
}
\description{

}
\details{

}
\examples{

}
\author{
Pascal Belleau
}
