% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseHLADataset.R
\name{readHLADataset}
\alias{readHLADataset}
\title{Extract HLA typing information for all samples present in a text file}
\usage{
readHLADataset(hlaFilePath)
}
\arguments{
\item{hlaFilePath}{a \code{character} string, the filepath to a text file
containing the information about the samples typing. The name must
correspond to an existing text file. The text file must have a header.}
}
\value{
a \code{list} of class \code{HLADataset}
containing the following elements:
\itemize{
\item \code{data} a \code{tibble} object containing the HLA typing
information for all samples. The columns are:
\itemize{
\item \code{SampleName} a \code{character} string that represent the
name of the sample.
\item \code{AllelName} a \code{character} string that represent the
name of the allele (1 or 2).
\item \code{GeneName} a \code{character} string that represent the
name of the HLA gene.
\item \code{AlleleGroup} a \code{character} string that represent the
section identifying the subtype group.
\item \code{Protein} a \code{character} string that represent the
section identifying the nucleotide substitutions group.
\item \code{SynSubst} a \code{character} string that represent the
section identifying the synonymous nucleotide substitutions group.
\item \code{NonCoding} a \code{character} string that represent the
section identifying the non-coding polymorphisms group.
\item \code{Suffix} a \code{character} string that represent the
suffix of the HLA typing or \code{NA}.
}
}
}
\description{
Parses a text file containing HLA allele typing for one or
multiple samples and transforms it into an object.
}
\examples{

## Get path where some HLA database files are stored
directory <- system.file("extdata", package = "HLAClustRView")
fileName <- paste0(directory, "/Samples_HLA_typing.txt")

## Parse file to extract HLA typing for all samples
ex <- readHLADataset(hlaFilePath=fileName)

## Show the output HLADataset object
print(ex)


}
\author{
Adewunmi Adelaja
}
