% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseHLADb.R
\name{readHLADataset}
\alias{readHLADataset}
\title{Extract HLA typing information for all samples present in a text file}
\usage{
readHLADataset(hlaDbPath)
}
\arguments{
\item{hlaDbPath}{a \code{character} string, the filepath to a text file
containing the information about the samples typing.}
}
\value{
a \code{data.frame} containing the following column names:
\itemize{
\item \code{SampleName} The name of the sample.
\item \code{AllelName} The name of the allele (1 or 2).
\item \code{GeneName} The name of the HLA gene.
\item \code{AlleleGroup} TODO
\item \code{Protein} TODO
\item \code{SynSubst} TODO
\item \code{NonCoding} TODO
\item \code{Suffix} The suffix of the HLA typing or \code{NA}.
}
}
\description{
Parses a text file containing HLA allele typing for one or
multiple samples and transforms it into an
HLADb object.
}
\examples{

## Get path where some HLA database files are stored
directory <- system.file("extdata", package = "HLAClustRView")
fileName <- paste0(directory, "/Samples_HLA_typing.txt")

## Parse file to extract HLA typing for all samples
ex <- readHLADataset(hlaDbPath=fileName)

## Show the first lines of the output dataset
head(ex)


}
\author{
Adewunmi Adelaja
}
